+++ src/axom/inlet/Inlet.cpp
@@ -23,33 +23,34 @@ namespace axom
 {
 namespace inlet
 {
-Table& Inlet::addStruct(const std::string& name, const std::string& description)
+Container& Inlet::addStruct(const std::string& name,
+                            const std::string& description)
 {
-  return m_globalTable.addStruct(name, description);
+  return m_globalContainer.addStruct(name, description);
 }
 
 VerifiableScalar& Inlet::addBool(const std::string& name,
                                  const std::string& description)
 {
-  return m_globalTable.addBool(name, description);
+  return m_globalContainer.addBool(name, description);
 }
 
 VerifiableScalar& Inlet::addDouble(const std::string& name,
                                    const std::string& description)
 {
-  return m_globalTable.addDouble(name, description);
+  return m_globalContainer.addDouble(name, description);
 }
 
 VerifiableScalar& Inlet::addInt(const std::string& name,
                                 const std::string& description)
 {
-  return m_globalTable.addInt(name, description);
+  return m_globalContainer.addInt(name, description);
 }
 
 VerifiableScalar& Inlet::addString(const std::string& name,
                                    const std::string& description)
 {
-  return m_globalTable.addString(name, description);
+  return m_globalContainer.addString(name, description);
 }
 
 void Inlet::registerWriter(std::unique_ptr<Writer> writer)
@@ -65,16 +66,16 @@ namespace detail
  * generation purposes
  * 
  * \param [inout] writer The Writer to use for documentation
- * \param [in] table The current table to write
+ * \param [in] container The current container to write
  *******************************************************************************
  */
-void writerHelper(Writer& writer, const Table& table)
+void writerHelper(Writer& writer, const Container& container)
 {
   // Use a pre-order traversal for readability
-  writer.documentTable(table);
-  for(const auto& sub_table_entry : table.getChildTables())
+  writer.documentContainer(container);
+  for(const auto& sub_container_entry : container.getChildContainers())
   {
-    writerHelper(writer, *sub_table_entry.second);
+    writerHelper(writer, *sub_container_entry.second);
   }
 }
 
@@ -84,12 +85,12 @@ void Inlet::writeDoc()
 {
   if(m_docEnabled)
   {
-    detail::writerHelper(*m_writer, m_globalTable);
+    detail::writerHelper(*m_writer, m_globalContainer);
     m_writer->finalize();
   }
 }
 
-bool Inlet::verify() const { return m_globalTable.verify(); }
+bool Inlet::verify() const { return m_globalContainer.verify(); }
 
 }  // end namespace inlet
 }  // end namespace axom